+++ docs/versionhistory.rst
@@ -7,9 +7,10 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 - Fixed a misleading ``ValueError`` in the context of DNS failures
   (`#815 <https://github.com/agronholm/anyio/issues/815>`_; PR by @graingert)
-- Allowed ``wait_socket_readable`` and ``wait_socket_writable`` to accept a socket
-  file descriptor (`#824 <https://github.com/agronholm/anyio/pull/824>`_)
-  (PR by @davidbrochart)
+- Added ``wait_readable`` and ``wait_writable`` functions that accept an object with a
+  ``.fileno()`` method or an integer handle, and deprecated ``wait_socket_readable``
+  and ``wait_socket_writable``.
+  (`#824 <https://github.com/agronholm/anyio/pull/824>`_) (PR by @davidbrochart)
 
 **4.6.2**